requirements.txt
streamlit>=1.32.0
torch>=2.0.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
PyMuPDF>=1.23.0
groq>=0.4.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
